Nonfunctional service skeleton
This adds the skeleton of a Dockerfile + Flask app to monitor
the resources in a K8s cluster it's deployed into and report
out on the status of each node (as to whether it has the minimum
resources available as specified by the request). Mostly just
jotting things down right now, it's in a nonfunctional state
for another commit or two, but verified at least that the
`update_node_list()` function works more or less as intended.