Return ok/notokay status on request
This adds a `/v1/resources` GET endpoint with optional `cpu`
and `ram` URL parameters. The endpoint will return an 'OK' status
for any nodes with capacity equal to or greater than that
specified in the request, or 'NOTOKAY' for any nodes with less
than that capacity on either criterion. I still need to do a
bit of testing to make sure that if I add/remove nodes to/from
the cluster while this service is running it'll pick them up
correctly, and need to test with heterogeneously provisioned
nodes, but then should be good to stub out a quick manifest
and get this deployed into the cluster.