Add better handling for resource quantities
Kubernetes uses "fun" SI siffices or their power-of-two equivalents
to represent resource quantities. This becomes tedious pretty quickly,
because literally who ever thinks of things in terms of kibibytes?

Nobody, that's who.

Anyway, this adds an extra function to convert from, e.g., 100Ki
to the equivalent 100 * 2^10 bytes and then into megabytes along
with the bare minimum error checking to make sure that the resource
value conforms to what I'm expecting.

Thankfully this is a nonissue for CPU since number of threads caps
out muuuuuuuch lower.