Deploy into cluster and port forward to local
This adds some additional stuff to build & run the image
with Skaffold, and adds port forwarding from the cluster
to your local machine. Verified that the service deploys
and is reachable on `localhost:8080`.

There's a little bit of RBAC thrown in here as well for
good measure because I forgot about it...